Class refactoring and cleanup.
