+++ tests/src/parser-test.js
@@ -1,7 +1,6 @@
 import {mathjax} from '../node_modules/mathjax-full/js/mathjax.js';
 
 import {TeX} from '../node_modules/mathjax-full/js/input/tex.js';
-import {SVG} from '../node_modules/mathjax-full/js/output/svg.js';
 import {RegisterHTMLHandler} from "../node_modules/mathjax-full/js/handlers/html.js";
 import {chooseAdaptor} from "../node_modules/mathjax-full/js/adaptors/chooseAdaptor.js";
 import {JsonMmlVisitor} from '../node_modules/mathjax-full/js/core/MmlTree/JsonMmlVisitor.js';
@@ -9,23 +8,27 @@ import {STATE} from '../node_modules/mathjax-full/js/core/MathItem.js';
 
 import {TagsFactory} from '../node_modules/mathjax-full/js/input/tex/Tags.js';
 import {MapHandler} from "../node_modules/mathjax-full/js/input/tex/MapHandler.js";
+import {AllPackages} from '../node_modules/mathjax-full/js/input/tex/AllPackages.js';
 
 import {JsonTest, Test} from './test.js';
-
-import {AllPackages} from '../node_modules/mathjax-full/js/input/tex/AllPackages.js';
-import '../node_modules/mathjax-full/js/input/tex/base/BaseConfiguration.js';
-import '../node_modules/mathjax-full/js/input/tex/ams/AmsConfiguration.js';
 
 let fs = require('fs');
 
 RegisterHTMLHandler(chooseAdaptor());
 
+
+/**
+ * Parser tests from JSON input files.
+ */
 export class ParserJsonTest extends JsonTest {
 
   packages = ['ams', 'base'];
   settings = {tags: 'none'};
   name = '';
 
+  /**
+   * @override
+   */
   constructor(json) {
     super(json);
     this.packages = this.json['packages'] || this.packages;
@@ -46,7 +49,15 @@ export class ParserJsonTest extends JsonTest {
     }
   }
 
+  document(options) {
+    return mathjax.document('<html></html>', {
+      InputJax: new TeX(options)
+    });
+  }
 
+  /**
+   * @override
+   */
   runTest(name, tex, expected, rest) {
     this.test(
       name,
@@ -54,10 +65,7 @@ export class ParserJsonTest extends JsonTest {
         mathjax.handleRetriesFor(function() {
           let options = {packages: this.packages};
           Object.assign(options, this.settings);
-          let html = mathjax.document('<html></html>', {
-            InputJax: new TeX(options), OutputJax: new SVG()
-          });
-          let root = html.convert(tex, {end: STATE.CONVERT});
+          let root = this.document(options).convert(tex, {end: STATE.CONVERT});
           let jv = new JsonMmlVisitor();
           root.setTeXclass(null);
           let actual = jv.visitTree(root);
@@ -71,3 +79,69 @@ export class ParserJsonTest extends JsonTest {
   }
 
 }
+
+
+import {SVG} from '../node_modules/mathjax-full/js/output/svg.js';
+
+/**
+ * Parser tests for packages that require an output jax.
+ */
+export class ParserOutputTest extends ParserJsonTest {
+
+  /**
+   * @override
+   */
+  document(options) {
+    return mathjax.document('<html></html>', {
+      InputJax: new TeX(options), OutputJax: new SVG()
+    });
+  }
+  
+}
+
+import {Configuration} from '../node_modules/mathjax-full/js/input/tex/Configuration.js';
+import {CharacterMap} from '../node_modules/mathjax-full/js/input/tex/SymbolMap.js';
+import ParseMethods from '../node_modules/mathjax-full/js/input/tex/ParseMethods.js';
+
+/**
+ * Parser tests for multi relations method.
+ */
+export class ParserMultirelTest extends ParserJsonTest {
+
+  constructor(json) {
+    // Just some things bogus attributes for testing.
+    new CharacterMap('shadow', ParseMethods.mathchar0mo, {
+      sim: ['\u223C', {something: 'nothing'}],
+      simeq: ['\u2243', {something: 'nothing'}],
+      asymp: ['\u224D', {something: 'else'}],
+      cong: ['\u224D', {anything: 'nothing'}],
+      lesssim: ['\u2272', {lspace: '1pt'}],
+      gtrsim: ['\u2278', {rspace: '1pt'}]
+    });
+    Configuration.create('multirel-test', {handler: {macro: ['shadow']}, priority: 4});
+    super(json);
+    this.packages = ['base', 'ams', 'multirel-test'];
+  }
+
+}
+
+
+/**
+ * Parser tests for macro configurations.
+ */
+export class ParserConfigMacrosTest extends ParserJsonTest {
+
+  constructor(json) {
+    super(json);
+    this.packages = ['base', 'configmacros'];
+    Object.assign(this.settings, {macros:{}});
+  }
+
+  runTest(name, input, expected, {macro, control}) {
+    Object.assign(this.settings.macros, macro);
+    super.runTest(name, control, expected);
+    super.runTest(name, input, expected);
+  }
+  
+}
+
